refactor(ledger): collocate the ledger route registrars :hammer:
Both ledger registrars move out of ledger_handler.go into ledger_routes.go,
leaving the handler file with the handler struct, the request/response
envelopes and the cores.

No route, authz tuple, operation ID or path changed.
